refactor(agent): extract agent tool dispatch
